agent/requirements.txt
pyyaml>=6.0
httpx>=0.26.0
psutil>=5.9.0
pytest>=7.4.0
pytest-httpx>=0.28.0
